+++ src/commands/hardis/org/files/import.ts
@@ -87,8 +87,8 @@ The command's technical implementation involves:
         type: 'confirm',
         name: 'value',
         message:
-          'Do you want to overwrite the existing files with the same name ?',
-        description: 'Replace existing files in Salesforce with local versions (doubles the number of API calls used)',
+          'Do you want to overwrite the existing files with the same name?',
+        description: 'Replace existing files in Salesforce with local versions (doubles the number of API calls used).',
       });
       this.handleOverwrite = handleOverwriteRes.value;
     }