+++ src/common/gitProvider/utilsMarkdown.ts
@@ -73,7 +73,7 @@ export function deployCodeCoverageToMarkdown(orgCoverage: number, orgCoverageTar
   if (orgCoverage < orgCoverageTarget) {
     return `❌ Your code coverage is insufficient: **${orgCoverage}%**, while your target is **${orgCoverageTarget}%**`;
   } else {
-    return `✅ Your code coverage is ok :) **${orgCoverage}%**, while target is **${orgCoverageTarget}%**`;
+    return `✅ Your code coverage is ok 😊 **${orgCoverage}%**, while target is **${orgCoverageTarget}%**`;
   }
 }
 
@@ -96,7 +96,7 @@ export async function flowDiffToMarkdownForPullRequest(flowNames: string[], from
     flowDiffFilesSummary += `- [${flowName}](#${flowName})\n`;
     const fileMetadata = await MetadataUtils.findMetaFileFromTypeAndName("Flow", flowName);
     try {
-      // Markdown with pure mermaidJs
+      // Markdown with pure MermaidJS
       if (supportsMermaidInPrMarkdown) {
         await generateDiffMarkdownWithMermaid(fileMetadata, fromCommit, toCommit, flowDiffMarkdownList, flowName);
       }